+++ src/components/js/inArray.js
