+++ src/components/js/removeProduct.js
