+++ src/components/js/totalCount.js
