+++ src/components/js/totalSum.js
