+++ src/components/jsx/Layout/layout/index.jsx
@@ -1,6 +1,6 @@
 import { Outlet } from "react-router-dom";
-import Header from "./header";
-import Footer from "./footer";
+import Header from "../header";
+import Footer from "../footer";
 
 function Layout() {
     return (
