+++ src/components/jsx/data/productsData/index.jsx
@@ -1,9 +1,9 @@
-import GetApi from "../api/getApi";
-import Loading from "./loading";
-import Error from "./error";
-import ProductsList from "./productsList";
-import { url } from "./js/constants";
-import useApi from "../store/apiStore";
+import GetApi from "../../api/getApi";
+import Loading from "../loading";
+import Error from "../error";
+import ProductsList from "../productsList";
+import { url } from "../../../js/constants";
+import useApi from "../../store/apiStore";
 import { shallow } from "zustand/shallow";
 
 function ProductsData() {
