+++ src/components/jsx/pages/checkoutSuccess/index.jsx
@@ -1,4 +1,4 @@
-import CheckoutButton from "../data/checkoutButton";
+import CheckoutButton from "../../data/checkoutButton";
 
 function CheckoutSuccess() {
