+++ src/components/jsx/styled/priceStyled/index.jsx
